pydobot-master/requirements.txt
pyserial
paho-mqtt
pydobot
flask
serialtools